+++ events/messageReactionRemove.js
@@ -1,0 +1,7 @@
+module.exports = async (client, reaction, user) => {
+    if (reaction.emoji.name !== process.env.RUSTY_EMOJI) return;
+    const message = reaction.message;
+    if (reaction.count >= 1) {
+        message.react(process.env.RUSTY_EMOJI);
+    }
+};
